refactoring: serviceWorker.js를 libs/... 으로 이동
