+++ frontend/guest-app/src/index.js
@@ -3,7 +3,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App/App.js";
-import * as serviceWorker from "./serviceWorker";
+import * as serviceWorker from "./libs/serviceWorker.js";
 
 ReactDOM.render(<App />, document.getElementById("root"));
