+++ frontend/guest-app/src/libs/serviceWorker.js
