+++ frontend/host-app/src/libs/serviceWorker.js
